Make use of mask and model individually set

+++ AT0206/Imaging/job_generator.py
@@ -96,7 +96,9 @@ mask_channel_name = \
 output_direc = os.path.join(data_directory, "single_channels/")
 
 # Use mask and model? Disable when continuing to clean.
-use_mask_model = True
+use_mask = True
+# Due to the current issues with the Arecibo data as a SD model, skip using it
+use_model = False
 
 # Run channels in given range
 channel_ms = []
@@ -144,10 +146,15 @@ for chan in channel_ms:
     # aren't there already
     if not os.path.exists(chan_ms):
         shutil.move(chan, channel_direc)
-    if not os.path.exists(model_name):
+    if not os.path.exists(model_name) and use_model:
         shutil.move(model_channel_name.format(mod_mask_num), channel_direc)
-    if not os.path.exists(mask_name):
+    if not os.path.exists(mask_name) and use_mask:
         shutil.move(mask_channel_name.format(mod_mask_num), channel_direc)
+
+    if not use_mask:
+        mask_name = "None"
+    if not use_model:
+        model_name = "None"
 
     chan_template = return_template(channel_direc, chan_ms,
                                     model_name, mask_name)
